moved autoencoders, encoders and decoders to NN.architectures package #50
